project: Removed `node_modules` directory.
There is no need to include `node_modules` folder in the project as  the node
modules are platform dependent. So if someone who tries to run windows
node_modules on linux or vice versa will be getting weird errors. Hence for
first time running the project on a new machine we do `npm install`.