project: Added `.gitignore`
`.gitignore` files are super important in a git based version controlled projects. It 
ignores the unnecessary files and clutter to be included in the repo. Also important
files which include `SECRETS` for environment variables, for eg db username pass
can also be excluded so that they aren't exposed to the public by mistake.
You can find common gitignore files by just googling the framework and appending gitignore to it. For example `NodeJs gitignore`.