config: Added a config directory and config file.
The need for this is to increase modularity and increase reusability. So if we are
using constants like `port number` multiple times in the project at different places
we do not need to search and change this data everwhere. Just changing it at one
place does the trick, reducing errors and manual grunt work.